Logger: Fix missing initial value; add automatic registration with the LoggingModule. (#204)
* Fix initial msg. Add automatic connection to the LoggingModule.

Now the Logger is a VariableGroup and in the prepare() method an empty
msg is written once. This makes sure the LoggingModule is not waiting
for initial values of Loggers and messages by other Loggers get lost.

* Add message in case on modules with Logger are found.

* Fix test and remove implementation of removed method.

* Do not use hard coded tags but use default tags.

Now all tags given to the LoggingModule are used to look for senders.