Remove StateMonitor as it makes no sense.
After discussion with Martin K. we concluded that this monitor makes no sense. Legit cases should use the ExactMonitor
instead (if e.g. a boolean on/off configuration switch is to be observed). The monitoring of a Status value makes
no sense, since we would then have a monitor monitoring another monitor. In that case, the behaviour of the other
monitor should either be corrected, or the monitors should be aggregated with the right aggregator mode.