TestFacility: do not wait until device backends are open
This was a left-over from times when no proper initial value propagation
and exception handling was implemented in ApplicationCore. Waiting until
the backends are open is anyway not sufficient, since the DeviceModule
needs to finish the recovery procedure. If any test would still rely on
this (anti-)feature, it would be subject to race conditions.

The wait is removed because it prevents important tests for the
exception handling specification. In particular, tests which
intentionally break devices during startup of the application would not
be implementable.