consistently update error status and message
Introduce StatusWithMessage and StatusWithMessageReader which handle this.
It is used in DeviceModule error reporting.
This is a preparation in order that we can report errors consistently to DOOCS, in DoocsAdapter

make StatusAggregator use StatusWithMessage

it should output always status code + string
on the input side, connects to either status or status + string
corresponding string input is found automatically
NOTE: DeviceModule API change, the status output variables are
renamed from (status, message) to (status, status_message)!

update testing of DeviceModule outputs

because of API change of DeviceModule  (renamed message ->
status_message)

fix crash and improve message and test

- fix crash due to incorrect handling of cast to Derived
- improve message output
- make existing tests pass
- add test for status message of StatusAggregator

move StatusWithMessageReaderBase and Status def to ControlSystemAdapter

since it is required from ApplicationCore as well as from the adapters