Let user write to alias via CS. Keep last message in the PV.
Before the message was always cleared after writing it. Now a new log
level is used to indicate the situation that write was called on the
message PV without calling sendMessage - it happens if the owning module
calls writeAll(). Now the last message of each sender is availble in the
CS.