Improve Logging module
* Create test case for testing multiple modules having the same.

* Use message IDs to support modules with similar names.

ToDo: Remove variable connections to CS like with MicroDAQ.

* Use automatic connection via CS. Rebuild hierarchy in Logging module.

* Put logging variables into a separate hierarchy level.

* Let user write to alias via CS. Keep last message in the PV.

Before the message was always cleared after writing it. Now a new log
level is used to indicate the situation that write was called on the
message PV without calling sendMessage - it happens if the owning module
calls writeAll(). Now the last message of each sender is availble in the
CS.

* Add test with modified hierarchy.

* Clang formaat applied.

* Hide internal PVs. Use different internal tag to work with regex.