fix exceptions in constructor not shown properly
They were previously hidden by the "BUG found" message in
ApplicationBase due to the missing call to shutdown(). This is now being
called automatically in the Application destructor if the application
was not yet started fully, in which case this is ok to do there
(calling shutdown before destroying the Application data members is
important only if threads have been started already).