Mac OS compatibility (#265)
* Use correct of version pthread_setname_np depending on platform.

On Linux, this function takes two parameters. On most other platforms
(including all BSD-based platforms) it only takes a single parameter.

* Omit -Wl,--export-dynamic flag on macOS.

This flag is not supported on Clang / LLVM and on macOS it is not needed
anway because exporting all symbols is the default behavior.

When using Clang on other platforms (e.g. Linux) --export_dynamic or
-rdynamic might have to be specified instead.

* clang compiler compatibility

Co-authored-by: Sebastian Marsching <sebastian.marsching-git-2016@aquenos.com>