fix unregisterModule() breaking triggers
A PV which is received by a module and also used as a trigger for a
DeviceModule was "deleted" from the model, if the module is unregistered
from its owner. This deletion is now prevented by adding a trigger
relationship to the model (which was anyway missing).

Note: the "deletion" never was a real deletion, instead only all
relationships are removed. As a consequence, the DeviceModule was still
pointing to the "deleted" vertex resulting in a wrong connection later.