fix: type collisions with constants
Constants (as generated by ApplicationModule::constant()) were causing a
logic_error if accessors of different types were using the same constant
(created with the same value and the same value type upon creation).
This is possible since the type passed to constant() does not have to be
the same as the type of the accessor.

This is now fixed by giving each constant a fully unique name, so every
accessor using a constant will be in its own little variable network.