fix!: compilation issue with clang 18
We were holding an std::vector of an incomplete type
(VariableNetworkNode) in a struct (ProcessVariableProperties), which
is not allowed. This fix changes it into an std::vector of shared_ptrs.
Since we are exposing this vector via ProcessVariableProxy::getNodes(),
this presents a breaking API change. Since this is used mostly
internally, this should not break much.