fix: Rework the ReverseRecoveryDecorator
Fixes various deadlocks with the LNM BitRange accessor, due to calling
read() while already in a read transfer.

Only decorate the write direction (as direct pass-through). For the read
direction, implement a simple async read scheme that can be triggered
from the DeviceManager when required. Also split the
ReverseRecoveryDecorator from the ExceptionHandlingDecorator and just
put the latter on top of the former